File and Folder Restructure
Moved files and folder structure to conform with conventional pattern.
