--- libmp2t/libmp2t.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "tsadptfd.h"
-#include "tsmetadata.h"
-#include "tspes.h"
-#include "tspmt.h"
-#include "tsprsr.h"
-#include "tspat.h"
-#include "tspckt.h"
-#include "tspacker.h"
-#include "tsnit.h"
