--- libmp2t/tstype.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-typedef uint8_t BYTE;
-typedef uint16_t UINT16;
-typedef uint32_t UINT32;
-typedef uint64_t UINT64;
