+++ mp2tp/include/mp2tp/tstype.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <cstdint>
+
+typedef uint8_t BYTE;
+typedef uint16_t UINT16;
+typedef uint32_t UINT32;
+typedef uint64_t UINT64;
