+++ mp2tp/src/tspacker.cpp
@@ -1,4 +1,4 @@
-#include "tspacker.h"
+#include <mp2tp/tspacker.h>
 
 #include <cassert>
 #include <vector>
